Create README.md
The purpose of this project report is to forecast the future one-year prediction of 2023 of the selected companies in the Indian stock market. In this, We are taking only the closing price of the stocks to predict the future price. The report aims to determine which model among ARIMA, Prophet, and LSTM is best suited for stock price analysis and prediction. The report is specifically designed to be useful for trading people who are interested in investing in the stock market.